requirements.txt
pandas==2.2.3
openpyxl==3.1.5
faker==30.8.2
pyyaml==6.0.2
streamlit==1.49.1
pytest==8.3.4

# AI (optional locally; CI will skip)
transformers==4.46.3
torch==2.2.2
